pyproject.toml
[tool.poetry]
name = "api-library"
version = "0.1.0"
description = ""
authors = ["Jorge Vasconcelos"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = "*"
python-dotenv = "*"
SQLAlchemy = "*"
SQLAlchemy-serializer = "*"
mysql-connector-python = "*"
python-multipart = "*"
pycryptodome = "*"
redis = "*"
alembic = "*"
opentelemetry-distro = "*"
opentelemetry-exporter-otlp = "*"
opentelemetry-instrumentation-logging = "*"
gunicorn = "*"
pydantic = "*"
pypika = "*"
psycopg2-binary = "*"


[tool.poetry.group.dev.dependencies]
Faker = "*"
pytest = "*"
httpx = "*"
ruff = "*"
allure-pytest = "*"


[tool.ruff]
line-length = 120
target-version = "py312"

select = [
    "B", # flake8-bugbear
    "BLE", # flake8-blind-except
    "C4", # comprehensions
#    "D", # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "ICN", # flake8-import-conventions
    "ISC", # flake8-implicit-str-concat
    "PGH", # pygrep-hooks
    "PLC", # pylint
    "PLE", # pylint
    "PLR", # pylint import style
    "PLW", # pylint
    "RET", # flake8-return
    "RUF", # ruff-specific rules
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    "TID", # flake8-tidy-imports
    "UP", # pyupgrade
    "W", # pycodestyle
    "YTT", # flake8-2020

]
ignore = [
    "SIM102",
    "RUF013"
]

fixable = ["A", "B", "C", "D", "E", "F", "I", "F401", "UP007"]
unfixable = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
